CR kumy (partial)
