+++ gk_prototype/emberjs/github-fastboot-example/DEV_run.sh
@@ -3,4 +3,4 @@
 
 # ember serve
 # ember serve --proxy https://api.geokretymap.org
-ember serve --proxy https://private-1b51bc-geokretsapiv2mocks.apiary-mock.com
+ember serve --proxy https://private-1b51bc-geokretyapimocks.apiary-mock.com/
