+++ gk_prototype/emberjs/github-fastboot-example/app/app.js
@@ -3,6 +3,9 @@ import Resolver from './resolver';
 import loadInitializers from 'ember-load-initializers';
 import config from './config/environment';
 
+// sets up Ember.Inflector
+import './models/geokrety-inflector-rules';
+
 const App = Application.extend({
   modulePrefix: config.modulePrefix,
   podModulePrefix: config.podModulePrefix,
