+++ gk_prototype/emberjs/github-fastboot-example/app/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <title>GithubFastbootExample</title>
+    <title>GeoKrety - EmberJS Prototype</title>
     <meta name="description" content="">
     <meta name="viewport" content="width=device-width, initial-scale=1">
