+++ gk_prototype/emberjs/github-fastboot-example/app/templates/about.hbs
@@ -2,6 +2,6 @@
   <div class="right tomster"></div>
   <h2>About</h2>
   <p>
-    <a href="https://www.emberjs.com/">EmberJS</a> prototype (by <a href="https://github.com/geokrety/">Geokrety</a> team)
+    <a href="https://www.emberjs.com/">EmberJS</a> prototype (by <a href="https://github.com/geokrety/">GeoKrety</a> team)
   </p>
 </div>
