+++ gk_prototype/emberjs/github-fastboot-example/app/templates/application.hbs
@@ -2,7 +2,7 @@
   <div class="menu">
     {{#link-to "index"}}
       <h1>
-        <em>GK Frontend</em>
+        <em>GeoKrety - EmberJS Prototype</em>
       </h1>
     {{/link-to}}
     <div class="links">
@@ -12,8 +12,8 @@
 	  {{#link-to "contact" class="menu-contact"}}
         Contact
       {{/link-to}}
-      {{#link-to "geokrets" class="menu-geokrets"}}
-        Geokrets
+      {{#link-to "geokrety" class="menu-geokrety"}}
+        GeoKrety
       {{/link-to}}
     </div>
   </div>
