+++ gk_prototype/emberjs/github-fastboot-example/app/templates/contact.hbs
@@ -1,6 +1,6 @@
 <div class="jumbo">
   <div class="right kretfoto"></div>
-  <h2>Geokrets contact</h2>
+  <h2>GeoKrety contact</h2>
   <p>
     Look at <a href="https://github.com/orgs/geokrety/people">geokrety/people</a>
   </p>
