+++ gk_prototype/emberjs/github-fastboot-example/app/templates/geokrety.hbs
@@ -1,7 +1,7 @@
 <!-- possible todo: use controller, view! -->
 <div class="jumbo">
   <div class="right"></div>
-  <h2>{{model.meta.count}} geokret{{#if (compare model.meta.count '>' 1)}}s{{/if}}</h2>
+  <h2>{{pluralize model.meta.count "GeoKret"}}</h2>
   <!-- component to show data -->
   {{#each model.data as |geokret|}}
 	{{geokret-list gk=geokret}}
