--- gk_prototype/emberjs/github-fastboot-example/app/templates/gitlabuser.hbs
@@ -1,13 +1,0 @@
-> gitlabuser.hbs
-
-<p>
-  Username: {{model.login}}
-</p>
-<p>
-  Avatar: <img src={{model.avatar_url}} width="32" height="32">
-</p>
-<p>
-  # of Public Repos: {{model.public_repos}}
-</p>
-
-{{outlet}}
