+++ gk_prototype/emberjs/github-fastboot-example/package.json
@@ -35,6 +35,7 @@
     "ember-data": "~3.0.0",
     "ember-export-application-global": "^2.0.0",
     "ember-fetch": "^3.4.4",
+    "ember-inflector": "^2.2.0",
     "ember-load-initializers": "^1.0.0",
     "ember-maybe-import-regenerator": "^0.1.6",
     "ember-resolver": "^4.0.0",
